+++ src/PSRule/Parser/RuleLexer.cs
@@ -63,7 +63,7 @@ namespace PSRule.Parser
         /// <summary>
         /// Read synopsis.
         /// </summary>
-        private bool Synopsis(TokenStream stream, RuleDocument doc)
+        private static bool Synopsis(TokenStream stream, RuleDocument doc)
         {
             if (!IsHeading(stream.Current, RULE_ENTRIES_HEADING_LEVEL, DocumentStrings.Synopsis))
                 return false;
@@ -76,7 +76,7 @@ namespace PSRule.Parser
         /// <summary>
         /// Read description.
         /// </summary>
-        private bool Description(TokenStream stream, RuleDocument doc)
+        private static bool Description(TokenStream stream, RuleDocument doc)
         {
             if (!IsHeading(stream.Current, RULE_ENTRIES_HEADING_LEVEL, DocumentStrings.Description))
                 return false;
@@ -89,7 +89,7 @@ namespace PSRule.Parser
         /// <summary>
         /// Read recommendation.
         /// </summary>
-        private bool Recommendation(TokenStream stream, RuleDocument doc)
+        private static bool Recommendation(TokenStream stream, RuleDocument doc)
         {
             if (!IsHeading(stream.Current, RULE_ENTRIES_HEADING_LEVEL, DocumentStrings.Recommendation))
                 return false;
@@ -102,7 +102,7 @@ namespace PSRule.Parser
         /// <summary>
         /// Read notes.
         /// </summary>
-        private bool Notes(TokenStream stream, RuleDocument doc)
+        private static bool Notes(TokenStream stream, RuleDocument doc)
         {
             if (!IsHeading(stream.Current, RULE_ENTRIES_HEADING_LEVEL, DocumentStrings.Notes))
                 return false;
@@ -115,7 +115,7 @@ namespace PSRule.Parser
         /// <summary>
         /// Read links.
         /// </summary>
-        private bool Links(TokenStream stream, RuleDocument doc)
+        private static bool Links(TokenStream stream, RuleDocument doc)
         {
             if (!IsHeading(stream.Current, RULE_ENTRIES_HEADING_LEVEL, DocumentStrings.Links))
                 return false;