+++ src/PSRule/Pipeline/PipelineWriter.cs
@@ -5,6 +5,7 @@ using PSRule.Configuration;
 using PSRule.Rules;
 using System.Collections.Generic;
 using System.Management.Automation;
+using System.Text;
 
 namespace PSRule.Pipeline
 {