+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   - Added character case assertion helpers `IsLower`, and `IsUpper`. [#555](https://github.com/microsoft/PSRule/issues/555)
     - `IsLower` checks that all letters in a field value are lowercase.
     - `IsUpper` checks that all letters in a field value are uppercase.
+- Bug fixes:
+  - Fixed NUnit report reasons should be escaped in markdown. [#471](https://github.com/microsoft/PSRule/issues/471)
 
 ## v0.21.0-B2009006 (pre-release)